backend/resume-parser/pyproject.toml
[project]
name = "simple-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.8",
    "langchain-openai>=1.1.7",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
]
